Change to reflect overall page style
Removed from the internal style sheet the font style for the id tag associate with the newsletter form. Also added the class btn to the button to give the radius border and letter weight, and added as style a margin-top of 5px to give some space to the email box.